+++ sdk/core/azure-core/test/ut/policy.cpp
@@ -7,6 +7,23 @@
 
 #include <vector>
 
+namespace {
+class NoOpPolicy : public Azure::Core::Http::HttpPolicy {
+public:
+  std::unique_ptr<Azure::Core::Http::HttpPolicy> Clone() const override
+  {
+    return std::make_unique<NoOpPolicy>(*this);
+  }
+
+  std::unique_ptr<Azure::Core::Http::RawResponse> Send(
+      Azure::Core::Context const&,
+      Azure::Core::Http::Request&,
+      Azure::Core::Http::NextHttpPolicy) const override
+  {
+  }
+};
+} // namespace
+
 TEST(Policy, throwWhenNoTransportPolicy)
 {
   // Construct pipeline without exception
@@ -43,4 +60,34 @@ TEST(Policy, throwWhenNoTransportPolicyMessage)
   {
     EXPECT_STREQ("Invalid pipeline. No transport policy found. Endless policy.", err.what());
   }
+}
+
+TEST(Policy, ValuePolicy)
+{
+  using namespace Azure::Core;
+  using namespace Azure::Core::Http;
+
+  ValuePolicyOptions options;
+  options.AddHeaderValues({{"HdrKey1", "HdrVal1"}, {"HdrKey2", "HdrVal0"}});
+  options.AddQueryValues({{"QryKey1", "QryVal0"}, {"QryKey2", "QryVal2"}});
+
+  options.AddHeaderValues({{"HdrKey2", "HdrVal2"}, {"HdrKey3", "HdrVal3"}});
+  options.AddQueryValues({{"QryKey1", "QryVal1"}, {"QryKey3", "QryVal3"}});
+
+  HttpPipeline pipeline({std::make_unique<ValuePolicy>(options), std::make_unique<NoOpPolicy>()});
+  Request request(HttpMethod::Get, Url("https:://www.example.com"));
+
+  pipeline.Send(GetApplicationContext(), request);
+
+  auto headers = request.GetHeaders();
+  auto queryParams = request.GetUrl().GetQueryParameters();
+
+  ASSERT_EQ(
+      headers,
+      decltype(headers)({{"HdrKey1", "HdrVal1"}, {"HdrKey2", "HdrVal2"}, {"HdrKey3", "HdrVal3"}}));
+
+  ASSERT_EQ(
+      queryParams,
+      decltype(queryParams)(
+          {{"QryKey1", "QryVal1"}, {"QryKey2", "QryVal2"}, {"QryKey3", "QryVal3"}}));
 }